switched to a new packaged-based plugin system
plugins are no longer methods of crabsort
this makes it possible for plugins to exist anywhere
on the path, as long as they are inside specially named
folders. This means that a 3rd party repo can offer a
crabsort plugin, which does not have to be packaged
with crabsort in order to work.